+++ spring/src/test/java/br/net/yurinogueira/springsales/domain/service/ProductServiceTest.java
@@ -8,7 +8,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -23,7 +22,7 @@ class ProductServiceTest {
         Product product = new Product();
         product.setName("Coca-Nola");
         product.setDescription("Refrigerante sem Gás");
-        product.setBasePrice(BigDecimal.valueOf(4.5));
+        product.setBasePrice(4.5);
 
         testService.save(product);
 
@@ -35,7 +34,7 @@ class ProductServiceTest {
         Product product = new Product();
         product.setName("Pepde");
         product.setDescription("Suco com Gás");
-        product.setBasePrice(BigDecimal.valueOf(4.0));
+        product.setBasePrice(4.0);
 
         testService.save(product);
 
@@ -57,10 +56,10 @@ class ProductServiceTest {
         Product secondProduct = new Product();
         firstProduct.setName("Feijoada sabor Batata");
         firstProduct.setDescription("Realmente possui sabor de batata");
-        firstProduct.setBasePrice(BigDecimal.valueOf(10.5));
+        firstProduct.setBasePrice(10.5);
         secondProduct.setName("Feijoada sabor Churrasco");
         firstProduct.setDescription("Não possui nenhum sabor de churrasco");
-        firstProduct.setBasePrice(BigDecimal.valueOf(6.5));
+        firstProduct.setBasePrice(6.5);
 
         testService.save(firstProduct);
         testService.save(secondProduct);
@@ -83,7 +82,7 @@ class ProductServiceTest {
         Product product = new Product();
         product.setName("Batatinhas Doces");
         product.setDescription("Batatinhas feitas com Açúcar");
-        product.setBasePrice(BigDecimal.valueOf(3.0));
+        product.setBasePrice(3.0);
 
         testService.save(product);
 
@@ -97,7 +96,7 @@ class ProductServiceTest {
         Product product = new Product();
         product.setName("Pipoca de Arroz");
         product.setDescription("Arroz em formato de pipoca");
-        product.setBasePrice(BigDecimal.valueOf(3.5));
+        product.setBasePrice(3.5);
 
         testService.save(product);
         testService.delete(product.getId());
